LPD-15362 Move CTCollectionHistoryDataProvider and remove unused methods
